requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.5
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
dateparser==1.1.0
fastapi==0.70.0
fastapi-utils==0.2.1
frozenlist==1.2.0
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
Mako==1.1.6
MarkupSafe==2.0.1
multidict==5.2.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.2
py==1.11.0
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.5
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==9.1
yarl==1.7.2
zipp==3.6.0
